requirements.txt
streamlit
pillow
ultralytics
roboflow